Check in code.
This is a basic Dockerfile and a perl wrapper for running
Cassandra inside a container. The perl wrapper takes the
existing cassandra.yaml from /etc/cassandra and rewrites
settings needed to make it work inside a container bound
to the right addresses, etc..

This pass uses the Datastax Community Edition, which is
the latest stable version of Apache Cassandra that has
gone through some QA and is packaged in a nice apt repo.