add symlink command support, ~/.cassandra, rename stress
Added symlink support so the entrypoint can be linked to the command
name to make it capture commands coming in via docker exec or on the
shell in debug sessions. e.g. `docker exec -it $ID cqlsh` is a symlink
from /bin/cassandra-docker to /bin/cqlsh. See also: Dockerfile, where
the symlinks are set up.

The entrypoint also creates /data/.cassandra and sets permissions on
it to keep cqlsh happy. It will probably also be handy to persist
the history.

Everyone knows cassandra-stress as cassandra-stress so leave it with
that name instead of calling it 'stress'.