bind by interface instead of IP
This should make it easier to reuse existing configs for things like
testing even if the IPs change. This commit has one caveat in that
broadcast_rpc_address is unset, Cassandra will NPE on startup. For
now it's still being set to the DefaultIP which should be fine for
most setups where the IP might change. It can be removed when
2.1.4 is out. https://issues.apache.org/jira/browse/CASSANDRA-8925